Fix #115 - Allow filters to be enabled / disabled per subscription
This adds a new field to filter state: `disabledSubscriptions`. This holds the set of subscription urls for which a given filter is disabled. For example, this allows having the same filter appear in both the user's custom filters and easylist, but disabling it in the user's custom filters does not also disable the filter in easylist.

The previous behaviour, enabling and disabling filters globally regardless of which subscriptions they're in, will continue to work as before to maintain backwards compatibility. However, this old behaviour is deprecated.

This also introduces a new event to the filter notifier: `filterState.disabledSubscriptions`, which emits an event whenever the list of subscriptions a filter is disabled for changes.

This also required a way to store sets in the serialized ini files.

I've also changed filterState to use a named class, since the JSDocs weren't rendering with the previous anonymous object syntax.