Refs #374 - Remove profiler option that doesn't exist anymore in Node 16
Frustratingly, there is still an option with the same name, but it does something completely different.

Unfortunately, this means that the profiler's callback is now always async. Benchmarking was relying on synchronous callbacks to ensure profiling data was available before writing results to file. I've replaced this with a function that checks if the expected data is in the results, and waits if it isn't there yet.

This was the only thing that was resulting in a failing test if you upgrade to NodeJS 16. After this change, it's viable for core developers to upgrade to NodeJS 16.