Fix #377 - Identify more types of filters which could be triggered by attribute changes
As a performance enhancement, ElemHideEmulation will only watch certain mutation events if it has filters which could be triggered by those events. One of the types of mutation events which can be filtered is attribute changes.

Unfortunately, some of the filters were being falsely flagged as unable to be triggered by attribute changes. This commit corrects that.