Fix #326 - Treat header and CSP separately from resource types
Header and CSP are currently being treated as fields that modify behaviour, as well as resource types. Unfortunately, the semantics that make sense for these two don't match other resource types.

If a filter is $header=foo,script, it should match only script requests that have the header foo. It should not also match stylesheets with the header foo. However, if a filter is JUST $header=foo, it should match both scripts and stylesheet.

To accommodate this, there are two major code changes

1. During parsing, consider $header and $csp as separate from the normal resource type flags. In other words, if a filter only specifies header, its content type is `contentTypes.HEADER | RESOURCE_TYPE` since the resource type flags are still defaulted to all.
2. When seeing if a filter matches, check for Header and CSP separately from other contentType flags. You need to match the Header / CSP flags AND the content type flag.

This would mean that consumers would need to always provide the type from the request as well as the contentTypes.HEADER or contentTypes.CSP context flags when doing header / csp matching.