Fix #392 - Return an allowing filter, even if it isn't overriding a blocking filter
Skipping allowing filters makes sense as a performance optimization in many situations. However, it does not work well in combination with header filtering. Now that we have header filters, an allowing filter might be relevant if it's found in the `onBeforeRequest` stage, because then you can skip any check on the `onHeadersReceived` stage.

This change allows core users to opt out of this optimization by passing the new `exhaustiveSearch` flag to `defaultMatcher.match`.