Fix #474 - Allow configuration of notification fetch intervals
Currently, notifications are fetched on a hardcoded schedule. It
starts 1 minute after the notifications module is started, and
the remote notification file fetched is considered to have
expired after 1 day. These values are generally fine as defaults
for core, but in webext-sdk we need to have the fetch start
sooner. This is because, when running in a service worker, the
service worker may be killed after 30 seconds of inactivity.

This commit makes the hardcoded timing values configurable through
Prefs. This is very similar to how timings for subscriptions
are also configurable through Prefs.