Fix #475 - Support saving local notifications between service worker restarts
By default, local notifications are stored in a Set. This commit
provides a mechanism for swapping that hashset for a different
object that has the same interface.

The webext-sdk project needs this functionality to support saving
local notifications to session storage, so that they can be
persisted between storage worker restarts.